@@ -2,7 +2,7 @@ name: radial_hero_animation_animate_rectclip
 description: Shows how to apply a radial transformation to the Hero as it animates to the new route.
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
