@@ -2,7 +2,7 @@ name: staggered_pic_selection
 description: A new Flutter project.
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
