@@ -3,12 +3,14 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 class BouncingBallDemo extends StatefulWidget {
+  @override
   _BouncingBallDemoState createState() => _BouncingBallDemoState();
 }
 
 class _BouncingBallDemoState extends State<BouncingBallDemo> with SingleTickerProviderStateMixin {
-  AnimationController controller;
+  late AnimationController controller;
 
+  @override
   void initState() {
     super.initState();
     controller = AnimationController(
@@ -17,23 +19,24 @@ class _BouncingBallDemoState extends State<BouncingBallDemo> with SingleTickerPr
       lowerBound: 0,
       upperBound: 100,
     );
-
   }
 
   @override
   Widget build(BuildContext context) {
     return Container(
       margin: EdgeInsets.only(top: 0),
-        child: Container(
-          decoration: BoxDecoration(
-            shape: BoxShape.circle,
-            color: Colors.green,
-          ),
-          width: 40.0,
-          height: 40.0,
-        )
-      );
+      child: Container(
+        decoration: BoxDecoration(
+          shape: BoxShape.circle,
+          color: Colors.green,
+        ),
+        width: 40.0,
+        height: 40.0,
+      ),
+    );
   }
+
+  @override
   void dispose() {
     controller.dispose();
     super.dispose();
