@@ -3,7 +3,7 @@ import 'package:test/test.dart';
 
 void main() {
   group('end-to-end test', () {
-    FlutterDriver driver;
+    late FlutterDriver driver;
 
     setUpAll(() async {
       // Connect to a running Flutter application instance.
@@ -11,7 +11,7 @@ void main() {
     });
 
     tearDownAll(() async {
-      if (driver != null) driver.close();
+      driver.close();
     });
 
     test('tap on the floating action button; verify counter', () async {
